Phase 3: address code-review findings
Correctness:
- Counter effect `by` is integer-only (rules.ts) — a fractional `by` no longer
  drives the cache non-integer and break counter/export reads. (#1)
- createRule rejects conditioning on the reserved counter_* types (a silent
  no-op: those events never reach the engine) and rejects custom ids in the
  default pack's R# namespace (so a future pack reseed can't be shadowed). (#4, #10)
- rebuildCounters now zeroes counters, clears the trace, and replays every event
  through the SAME apply-path used live (applyDirectManipulation + applyRules),
  so rebuilt values AND trace match live application by construction and every
  rebuilt change has a recorded cause. Documents that replay uses the current
  rule set (rule effective-dating is a later phase). Removes the duplicated
  fired-rule→counter fold. (#2, #3, #8)
- ensureRulePackSeeded upserts the definition on a version bump (ON CONFLICT DO
  UPDATE) so a corrected default rule/counter actually reaches already-seeded
  couples; live counter value and a rule's enabled toggle are preserved. (#7)

Noise / clarity:
- Near-misses are recorded only when a rule is pending on one of the event
  type's `awaiting` keys (threaded through RuleEventContext.awaiting), so routine
  events don't bury the trace in didn't-fire lines. (#6)
- Chain-view header is the neutral "Effects & near-misses" so a manual +1/reset
  tap isn't misattributed to the rules engine. (#5)
- resolveMatchOn documents the Phase-4 contract for an unresolved match key
  (treat as "no matching timer", never "match all"). (#9)

Tests: 52 passing (added awaiting-gated near-miss + fractional-by rejection).

+++ src/shared/rules.ts
@@ -26,12 +26,14 @@ export const effectSchema = z.discriminatedUnion("verb", [
 	z.object({
 		verb: z.literal("increment_counter"),
 		counter: z.string(),
-		by: z.number().default(1),
+		// Integer only — counter values are integers (counterSchema.value.int()); a
+		// fractional `by` would drive the cache non-integer and break reads/export.
+		by: z.number().int().default(1),
 	}),
 	z.object({
 		verb: z.literal("decrement_counter"),
 		counter: z.string(),
-		by: z.number().default(1),
+		by: z.number().int().default(1),
 	}),
 	z.object({ verb: z.literal("reset_counter"), counter: z.string() }),
 	z.object({ verb: z.literal("reset_anchor"), anchor: z.string() }),
